batchhandler: move flagging handlers/support code
Flagging of issues is complex enough to warrant its own file now
